requirements.txt
# Launcher Requirements
streamlit
psutil
streamlit-option-menu
watchdog

# Submodule Requirements
yfinance
pandas
curl_cffi
deepdiff
pillow
pytesseract
google-generativeai
pytest
libsql-client
altair
requests
urllib3
pytz
lightweight-charts
numpy
streamlit-js-eval
beautifulsoup4
feedparser
textblob
plotly
cloudscraper
selenium
webdriver-manager
investpy
infisical-python

# Note: 'ecocal' was excluded as it is not available on PyPI
streamlit-option-menu
infisical-python
streamlit-lightweight-charts
